Moved the interfaces into models folder and added table for features
