+++ src/app/_helpers/authenticationLogin.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpRequest, HttpResponse, HttpHandler, HttpEvent, HttpInterceptor, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Observable, of, throwError } from 'rxjs';
 import { delay, mergeMap, materialize, dematerialize } from 'rxjs/operators';
-import { User } from "../Shared/user";
+import { User } from "../models/user";
 import { USER } from "../Shared/users";
 
 let users = USER;
