+++ src/app/features/features.component.html
@@ -1,3 +1,29 @@
-<p>
-  features works!
-</p>
+<div class="container">
+  <h2>Features List</h2>
+  <p>Following are the list of features available that can we used to show/hide sections on the portal or it can be used at the service side as well.</p>
+  <div class="table-responsive">
+    <table class="table table-hover">
+      <thead>
+        <tr>
+          <th>#</th>
+          <th>Name</th>
+          <th>Description</th>
+          <th>Created Date</th>
+          <th>Last Updated</th>
+          <th>Toggle</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td>1</td>
+          <td>Anna</td>
+          <td>Pitt</td>
+          <td>35</td>
+          <td>New York</td>
+          <td>USA</td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
+</div>
+
