+++ src/app/home/home.component.ts
@@ -5,9 +5,9 @@ import { Promotion } from '../shared/promotion';
 import { PromotionService } from '../services/promotion.service';
 import {Leader} from '../shared/leader';
 import { LeaderService } from "../services/leader.service";
-import { Course } from "../Shared/course";
+import { Course } from "../models/course";
 import { COURSES } from "../Shared/courses";
-import { Feature } from "../Shared/feature";
+import { Feature } from "../models/feature";
 import { FEATURES } from "../Shared/features";
 import { AuthenticationService } from "../services/authentication.service";
 import { BehaviorSubject, Observable } from 'rxjs';
